requirements.txt
colorama==0.4.6
coverage==7.6.9
iniconfig==2.0.0
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
pytest==8.3.4
pytest-cov==6.0.0
scipy==1.14.1
setuptools==75.6.0

